# swagger.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Node Backend Boilerplate
  description: NodeJs backend
  termsOfService: https://github.com/flavien-perier/nodeBackendBoilerplate/blob/master/LICENSE
  contact:
    email: perier@flavien.io
  license:
    name: MIT
    url: https://github.com/flavien-perier/nodeBackendBoilerplate/blob/master/LICENSE

externalDocs:
  description: Readme
  url: https://github.com/flavien-perier/nodeBackendBoilerplate/blob/master/README.md

servers:
  - url: http://127.0.0.1:8080/

tags:
  - name: Account
    description: User management
  - name: Api
    description: Application

paths:
  /account:
    post:
      tags:
        - Account
      summary: Create an account
      requestBody:
        description: User information on account creation
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInformation'
        "400":
          description: Bad request
        "403":
          description: Forbidden
          content: {}

  /account/login:
    get:
      tags:
        - Account
      summary: Authentification
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
        - client: []

  /ping:
    get:
      tags:
        - Api
      summary: Simple test function
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - ping
                properties:
                  ping:
                    type: string
                    enum:
                      - OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  /sum:
    get:
      tags:
        - Api
      summary: Sum 2 values
      parameters:
        - name: value1
          in: query
          description: Value 1
          required: true
          schema:
            type: integer
            format: int64
        - name: value2
          in: query
          description: Value 2
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - result
                properties:
                  result:
                    type: number
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      security:
        - bearer: []

components:
  schemas:
    Account:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
          description: Account name
        password:
          type: string
          description: Account password

    AccountInformation:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Account name

    Token:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Account authentication token

  securitySchemes:
    client:
      type: http
      scheme: basic

    bearer:
      type: http
      scheme: bearer